+++ ninja-core/src/main/java/ninja/session/Session.java
@@ -23,13 +23,19 @@ import ninja.Result;
 
 import com.google.inject.ImplementedBy;
 
-/**
- * Session Cookie... Mostly an adaption of Play1's excellent cookie system that
- * in turn is based on the new client side rails cookies.
- */
+
 @ImplementedBy(SessionImpl.class)
 public interface Session {
-	
+    
+    String AUTHENTICITY_KEY = "___AT";
+    String ID_KEY = "___ID";
+    String TIMESTAMP_KEY = "___TS";
+    
+	/**
+     * Has to be called initially. => maybe in the future as assisted inject.
+     * 
+     * @param context The context of this session.
+     */
 	public void init(Context context);
 
 	/**
@@ -38,40 +44,61 @@ public interface Session {
 	public String getId();
 
 	/**
-	 * @return complete content of session.
+	 * @return complete content of session as immutable copy.
 	 */
 	public Map<String, String> getData();
 
 	/**
-	 * @return an authenticity token or generates a new one.
+	 * @return a authenticity token (may generate a new one if the session
+     *         currently does not contain the token). 
 	 */
 	public String getAuthenticityToken();
 
+    /**
+     * To finally send this session to the user this method has to be called.
+     * It basically serializes the session into the header of the response.
+     * 
+     * @param context The context from where to deduct a potentially existing session.
+     * @param result The result where to add the session.
+     */
 	public void save(Context context, Result result);
 
-
+    /**
+     * Puts key / value into the session. 
+     * PLEASE NOTICE: If value == null the key will be removed!
+     * 
+     * @param key Name of the key to store in the session.
+     * @param value The value to store in the session
+     */
 	public void put(String key, String value);
 
 	/**
 	 * Returns the value of the key or null.
 	 * 
-	 * @param key
-	 * @return
+	 * @param key Name of the key to retrieve.
+	 * @return The value of the key or null.
 	 */
 	public String get(String key);
 
 	/**
 	 * Removes the value of the key and returns the value or null.
 	 * 
-	 * @param key
-	 * @return
+	 * @param key name of the key to remove
+	 * @return original value of the key we just removed
 	 */
 	public String remove(String key);
 
+    /**
+     * Removes all values from the session.
+     */
 	public void clear();
+
 	/**
 	 * Returns true if the session is empty, e.g. does not contain anything else
 	 * than the timestamp key.
+     * 
+     * @return true if session does not contain any values / false if it contains
+     *         values.
 	 */
 	public boolean isEmpty();