+++ ninja-core/src/main/java/ninja/session/SessionImpl.java
@@ -16,6 +16,7 @@
 
 package ninja.session;
 
+import com.google.common.collect.ImmutableMap;
 import java.io.UnsupportedEncodingException;
 import java.util.HashMap;
 import java.util.Map;
@@ -25,6 +26,7 @@ import ninja.Context;
 import ninja.Cookie;
 import ninja.Result;
 import ninja.utils.CookieDataCodec;
+import ninja.utils.CookieEncryption;
 import ninja.utils.Crypto;
 import ninja.utils.NinjaConstant;
 import ninja.utils.NinjaProperties;
@@ -34,35 +36,31 @@ import org.slf4j.LoggerFactory;
 
 import com.google.inject.Inject;
 
-/**
- * Session Cookie... Mostly an adaption of Play1's excellent cookie system that
- * in turn is based on the new client side rails cookies.
- */
+
 public class SessionImpl implements Session {
     
-    private static Logger logger = LoggerFactory.getLogger(SessionImpl.class);
-
-    private static final String AUTHENTICITY_KEY = "___AT";
-    private static final String ID_KEY = "___ID";
-    private static final String TIMESTAMP_KEY = "___TS";
+    private final static Logger logger = LoggerFactory.getLogger(SessionImpl.class);
 
     private final Crypto crypto;
+    private final CookieEncryption encryption;
 
     private final Integer sessionExpireTimeInMs;
     private final Boolean sessionSendOnlyIfChanged;
     private final Boolean sessionTransferredOverHttpsOnly;
     private final Boolean sessionHttpOnly;
-    private final String applicationCookiePrefix;
     private final String applicationCookieDomain;
     private final Map<String, String> data = new HashMap<String, String>();
 
     /** Has cookie been changed => only send new cookie stuff has been changed */
     private boolean sessionDataHasBeenChanged = false;
+    
+    private final String sessionCookieName;
 
     @Inject
-    public SessionImpl(Crypto crypto, NinjaProperties ninjaProperties) {
+    public SessionImpl(Crypto crypto, CookieEncryption encryption, NinjaProperties ninjaProperties) {
 
         this.crypto = crypto;
+        this.encryption = encryption;
 
         // read configuration stuff:
         Integer sessionExpireTimeInSeconds = ninjaProperties
@@ -81,44 +79,36 @@ public class SessionImpl implements Session {
         this.sessionHttpOnly = ninjaProperties.getBooleanWithDefault(
                 NinjaConstant.sessionHttpOnly, true);
 
-        this.applicationCookiePrefix = ninjaProperties
-                .getOrDie(NinjaConstant.applicationCookiePrefix);
-
         this.applicationCookieDomain = ninjaProperties
                 .get(NinjaConstant.applicationCookieDomain);
+        
+        String applicationCookiePrefix = ninjaProperties
+                .getOrDie(NinjaConstant.applicationCookiePrefix);
+        this.sessionCookieName = applicationCookiePrefix + ninja.utils.NinjaConstant.SESSION_SUFFIX;
     }
 
-    /**
-     * Has to be called initially. => maybe in the future as assisted inject.
-     * 
-     * @param context
-     */
     @Override
     public void init(Context context) {
 
         try {
 
             // get the cookie that contains session information:
-            Cookie cookie = context.getCookie(applicationCookiePrefix
-                    + ninja.utils.NinjaConstant.SESSION_SUFFIX);
+            Cookie cookie = context.getCookie(sessionCookieName);
 
             // check that the cookie is not empty:
-            if (cookie != null && cookie.getValue() != null
-                    && !cookie.getValue().trim().equals("")) {
+            if (cookie != null && cookie.getValue() != null && !cookie.getValue().trim().isEmpty()) {
 
                 String value = cookie.getValue();
 
                 // the first substring until "-" is the sign
                 String sign = value.substring(0, value.indexOf("-"));
 
-                // rest from "-" until the end it the payload of the cookie
+                // rest from "-" until the end is the payload of the cookie
                 String payload = value.substring(value.indexOf("-") + 1);
 
                 // check if payload is valid:
-                // if (sign.equals(crypto.signHmacSha1(payload))) {
-
-                if (CookieDataCodec.safeEquals(sign,
-                        crypto.signHmacSha1(payload))) {
+                if (CookieDataCodec.safeEquals(sign, crypto.signHmacSha1(payload))) {
+                    payload = encryption.decrypt(payload);
                     CookieDataCodec.decode(data, payload);
                 }
 
@@ -149,9 +139,6 @@ public class SessionImpl implements Session {
         }
     }
 
-    /**
-     * @return id of a session.
-     */
     @Override
     public String getId() {
         if (!data.containsKey(ID_KEY)) {
@@ -161,17 +148,11 @@ public class SessionImpl implements Session {
 
     }
 
-    /**
-     * @return complete content of session.
-     */
     @Override
     public Map<String, String> getData() {
-        return data;
+        return ImmutableMap.copyOf(data);
     }
 
-    /**
-     * @return an authenticity token or generates a new one.
-     */
     @Override
     public String getAuthenticityToken() {
         if (!data.containsKey(AUTHENTICITY_KEY)) {
@@ -183,8 +164,7 @@ public class SessionImpl implements Session {
     @Override
     public void save(Context context, Result result) {
 
-        // Don't save the cookie nothing has changed, and if we're not expiring
-        // or
+        // Don't save the cookie nothing has changed, and if we're not expiring or
         // we are expiring but we're only updating if the session changes
         if (!sessionDataHasBeenChanged
                 && (sessionExpireTimeInMs == null || sessionSendOnlyIfChanged)) {
@@ -196,13 +176,10 @@ public class SessionImpl implements Session {
         if (isEmpty()) {
             // It is empty, but there was a session coming in, therefore clear
             // it
-            if (context.hasCookie(applicationCookiePrefix
-                    + NinjaConstant.SESSION_SUFFIX)) {
+            if (context.hasCookie(sessionCookieName)) {
 
-                Cookie.Builder expiredSessionCookie = Cookie.builder(
-                        applicationCookiePrefix + NinjaConstant.SESSION_SUFFIX,
-                        "");
-                expiredSessionCookie.setPath("/");
+                Cookie.Builder expiredSessionCookie = Cookie.builder(sessionCookieName, "");
+                expiredSessionCookie.setPath(context.getContextPath() + "/");
                 expiredSessionCookie.setMaxAge(0);
 
                 result.addCookie(expiredSessionCookie.build());
@@ -219,12 +196,14 @@ public class SessionImpl implements Session {
 
         try {
             String sessionData = CookieDataCodec.encode(data);
+            // first encrypt data and then generate HMAC from encrypted data
+            // http://crypto.stackexchange.com/questions/202/should-we-mac-then-encrypt-or-encrypt-then-mac
+            sessionData = encryption.encrypt(sessionData);
 
             String sign = crypto.signHmacSha1(sessionData);
 
-            Cookie.Builder cookie = Cookie.builder(applicationCookiePrefix
-                    + NinjaConstant.SESSION_SUFFIX, sign + "-" + sessionData);
-            cookie.setPath("/");
+            Cookie.Builder cookie = Cookie.builder(sessionCookieName, sign + "-" + sessionData);
+            cookie.setPath(context.getContextPath() + "/");
 
             if(applicationCookieDomain != null){
                 cookie.setDomain(applicationCookieDomain);
@@ -244,17 +223,11 @@ public class SessionImpl implements Session {
 
         } catch (UnsupportedEncodingException unsupportedEncodingException) {
             logger.error("Encoding exception - this must not happen", unsupportedEncodingException);
+            throw new RuntimeException(unsupportedEncodingException);
         }
 
     }
 
-    /**
-     * Puts key into session. PLEASE NOTICE: If value == null the key will be
-     * removed!
-     * 
-     * @param key
-     * @param value
-     */
     @Override
     public void put(String key, String value) {
 
@@ -274,12 +247,6 @@ public class SessionImpl implements Session {
 
     }
 
-    /**
-     * Returns the value of the key or null.
-     * 
-     * @param key
-     * @return
-     */
     @Override
     public String get(String key) {
         return data.get(key);
@@ -287,7 +254,6 @@ public class SessionImpl implements Session {
 
     @Override
     public String remove(String key) {
-
         sessionDataHasBeenChanged = true;
         String result = get(key);
         data.remove(key);
@@ -300,10 +266,6 @@ public class SessionImpl implements Session {
         data.clear();
     }
 
-    /**
-     * Returns true if the session is empty, e.g. does not contain anything else
-     * than the timestamp key.
-     */
     @Override
     public boolean isEmpty() {
         return (data.isEmpty() || data.size() == 1