+++ ninja-core/src/test/java/ninja/cache/NinjaCacheTest.java
@@ -16,16 +16,17 @@
 
 package ninja.cache;
 
-import net.sf.ehcache.CacheException;
-import ninja.utils.TimeUtil;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+import net.sf.ehcache.CacheException;
+import ninja.utils.TimeUtil;
+
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 import org.mockito.runners.MockitoJUnitRunner;
 
 @RunWith(MockitoJUnitRunner.class)