+++ ninja-core/src/test/java/ninja/utils/TimeUtilTest.java
@@ -16,7 +16,8 @@
 
 package ninja.utils;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;