+++ ninja-servlet/src/main/java/ninja/servlet/ContextImpl.java
@@ -59,9 +59,9 @@ import org.apache.commons.fileupload.FileUploadException;
 import org.apache.commons.fileupload.servlet.ServletFileUpload;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import com.google.inject.Inject;
-import org.slf4j.LoggerFactory;
 
 public class ContextImpl implements Context.Impl {